+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Changed
+- Form instantiation now recommended with form Element instead of form selector.
+
 [5.14.1] - 2020-03-31
 ----------------------
 ##### Fixed